Convert to partial memory representation.
As per our meeting on 2012-02-03, we converted the memory representation
in IL.v from total (W -> B) to partial (W -> option B).  We modified
the ReadWord primitive to be partial, and fixed up all of the main Bedrock
theorems that broke in the process.  Overall, this translation was
fairly pleasant.

This patch is incomplete, it does not:

    - Remove the now irrelevant inBounds checking
    - Add new predicates for properly asserting that a memory is
      appropriately total.  We may just want to do exists and ===>
    - Figure out why the 'ho' tactic stopped working